+++ packages/element-templates-json-schema-shared/test/helpers/index.js
@@ -1,17 +1,10 @@
-const {
-  forEach,
-  set
-} = require('min-dash');
-
-const chai = require('chai');
-
 const { default: Ajv } = require('ajv');
 const AjvErrors = require('ajv-errors');
 
+const { withErrorMessages, withDeprecationWarnings, getDeprecationWarnings } = require('./utils');
+
 module.exports = {
-  createValidator,
-  withErrorMessages,
-  withDeprecationWarnings
+  createValidator
 };
 
 function createValidator(schema, errors, deprecations) {
@@ -57,13 +50,6 @@ function createValidator(schema, errors, deprecations) {
 
   const validator = ajv.compile(withErrorMessages(withDeprecationWarnings(schema, deprecations), errors));
 
-
-  const getDeprecationWarnings = function(data) {
-    if (deprecationWarnings.length > 0 && data.id === deprecationWarnings[0].id) {
-      return deprecationWarnings.map(warning => warning.warningDescription);
-    }
-  };
-
   // wrapper function checks for warnings before each validation
   const wrappedValidator = function(data, ...args) {
 
@@ -73,87 +59,10 @@ function createValidator(schema, errors, deprecations) {
     const result = validator.call(this, data, ...args);
 
     wrappedValidator.errors = validator.errors;
-    wrappedValidator.warnings = getDeprecationWarnings(data);
+    wrappedValidator.warnings = getDeprecationWarnings(deprecationWarnings, data);
 
     return result;
   };
 
   return wrappedValidator;
-}
-
-function withErrorMessages(schema, errors) {
-
-  if (!errors || !errors.length) {
-    return schema;
-  }
-
-  // clone a new copy
-  let newSchema = JSON.parse(JSON.stringify(schema));
-
-  // set <errorMessage> keyword for given path
-  forEach(errors, function(error) {
-    newSchema = setErrorMessage(newSchema, error);
-  });
-
-  return newSchema;
-}
-
-function setErrorMessage(schema, error) {
-  const {
-    path,
-    errorMessage
-  } = error;
-
-  const errorMessagePath = [
-    ...path,
-    'errorMessage'
-  ];
-
-  return set(schema, errorMessagePath, errorMessage);
-}
-
-function eqlErrors(chai, utils) {
-
-  const Assertion = chai.Assertion;
-
-  Assertion.addMethod('eqlErrors', function(expectedErrors, filter) {
-
-    const actualErrors = this._obj;
-
-    // formats the validation errors, so that they can be used directly in the fixture files.
-    this.eql(expectedErrors,
-      `Errors from validation do not match expected.\n\tValidation returned this error (you can use it in the fixture):\n\t${JSON.stringify(actualErrors, null, 2).replace(/"([^"]+)":/g, '$1:')}\n`);
-  });
-}
-
-chai.use(eqlErrors);
-
-function withDeprecationWarnings(schema, deprecations) {
-  if (!deprecations || !deprecations.length) {
-    return schema;
-  }
-
-  // clone a new copy
-  let newSchema = JSON.parse(JSON.stringify(schema));
-
-  // set deprecation warnings for given paths
-  forEach(deprecations, function(deprecation) {
-    newSchema = setDeprecationWarning(newSchema, deprecation);
-  });
-
-  return newSchema;
-}
-
-function setDeprecationWarning(schema, deprecation) {
-  const {
-    path,
-    warningMessage
-  } = deprecation;
-
-  const deprecationPath = [
-    ...path,
-    'deprecatedWarning'
-  ];
-
-  return set(schema, deprecationPath, warningMessage);
 }
